testing_requirements.txt
beautifulsoup4>=4.0
pytest>=2.8
pytest-django>=2.9
mock
